Move WazuhClusterClient and check_cluster_status to cluster.py
